Project 4 - Dockerising multiple applications/web_app/requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
psycopg2-binary==2.9.5

